--- Matema/celery.py
@@ -1,12 +1,0 @@
-from __future__ import absolute_import, unicode_literals
-import os
-from celery import Celery
-
-
-os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'Matema.settings')
-
-app = Celery('Matema')
-
-app.config_from_object('django.conf:settings', namespace='CELERY')
-
-app.autodiscover_tasks()
