+++ Matema/scheduler.py
@@ -1,7 +1,3 @@
-
-# *************         WITHOUT CELERY AND CELERY BEAT          *************
-
-
 from apscheduler.schedulers.background import BackgroundScheduler
 from django_apscheduler.jobstores import DjangoJobStore, register_events, register_job
 from datetime import datetime
