+++ User/urls.py
@@ -1,7 +1,6 @@
 from django.urls import path
-from User.views import UserPhotoView, \
+from User.views import UserPhotoView, check_user_active, \
     UserScoreView, UserGetTaskView
-from User.service import check_user_active
 
 urlpatterns = [
