--- docker-compose.yml
@@ -1,19 +1,0 @@
-version: '3.8'
-
-services:
-  web:
-    build: .
-    volumes:
-      - .:/usr/src/app
-    ports:
-      - "8000:8000"
-    depends_on:
-      - redis
-    environment:
-      - REDIS_HOST=redis
-
-  redis:
-    image: redis:latest
-    ports:
-      - "6379:6379"
-
