--- fixtures/typeanswer_fixture.json
@@ -1,23 +1,0 @@
-[
-  {
-    "model": "Task.typeanswer",
-    "pk": 1,
-    "fields": {
-      "name": "short"
-    }
-  },
-  {
-    "model": "Task.typeanswer",
-    "pk": 2,
-    "fields": {
-      "name": "mcq"
-    }
-  },
-  {
-    "model": "Task.typeanswer",
-    "pk": 3,
-    "fields": {
-      "name": "maching"
-    }
-  }
-]
