--- fixtures/user_fixture.json
@@ -1,178 +1,0 @@
-[
-  {
-    "model": "User.user",
-    "pk": 2,
-    "fields": {
-      "password": "pbkdf2_sha256$600000$Te3TzeffRojw1PKttRHL5B$vEPxMXkj1pFHML8/m51llKb0RiT7y94tOc8TN1eO/eo=",
-      "last_login": "2023-12-18T17:21:28.189Z",
-      "is_superuser": true,
-      "username": "admin",
-      "first_name": "admin",
-      "last_name": "admin",
-      "email": "admin@gmail.com",
-      "is_staff": true,
-      "is_active": true,
-      "date_joined": "2023-12-11T12:39:14.534Z",
-      "image_url": "None",
-      "score": 0,
-      "count_tasks": 0,
-      "photo": "None",
-      "groups": [],
-      "user_permissions": []
-    }
-  },
-  {
-    "model": "User.user",
-    "pk": 5,
-    "fields": {
-      "password": "pbkdf2_sha256$600000$EoLRBGImPMdqXY9joVIbcl$oDX4I6Mumr8+dVzAB5I6HK+AI41xhuT9TEvYUFx9Cdg=",
-      "last_login": null,
-      "is_superuser": false,
-      "username": "MarikLNU_matema",
-      "first_name": "Maria",
-      "last_name": "Maria1",
-      "email": "matema.group@gmail.com",
-      "is_staff": false,
-      "is_active": false,
-      "date_joined": "2023-12-17T10:29:02.427Z",
-      "image_url": "None",
-      "score": 0,
-      "count_tasks": 0,
-      "photo": "None",
-      "groups": [],
-      "user_permissions": []
-    }
-  },
-  {
-    "model": "User.user",
-    "pk": 6,
-    "fields": {
-      "password": "pbkdf2_sha256$600000$5EVeYEzHTBTiPn9INVitS3$LL+SdFU58mJc/ZmMpSmQUiYpdwWIqE+tYA4fYnzVv14=",
-      "last_login": null,
-      "is_superuser": false,
-      "username": "ivanka4",
-      "first_name": "ivanna",
-      "last_name": "kozak",
-      "email": "ivanna@lpnu.ua",
-      "is_staff": false,
-      "is_active": false,
-      "date_joined": "2023-12-17T10:41:31.209Z",
-      "image_url": "None",
-      "score": 0,
-      "count_tasks": 0,
-      "photo": "None",
-      "groups": [],
-      "user_permissions": []
-    }
-  },
-  {
-    "model": "User.user",
-    "pk": 4,
-    "fields": {
-      "password": "pbkdf2_sha256$600000$r1992DWIYoG0shiMibXnlc$yrU9CoCt2u5Sb1mso9HXeoZm3J4AjGVw8QEh6m+7e+w=",
-      "last_login": null,
-      "is_superuser": false,
-      "username": "--hv--",
-      "first_name": "Матвій",
-      "last_name": "Готович",
-      "email": "matvii.hotovych.shi.2022@lpnu.ua",
-      "is_staff": false,
-      "is_active": true,
-      "date_joined": "2023-12-16T14:01:19Z",
-      "image_url": "None",
-      "score": 123,
-      "count_tasks": 0,
-      "photo": "",
-      "groups": [],
-      "user_permissions": []
-    }
-  },
-  {
-    "model": "User.user",
-    "pk": 7,
-    "fields": {
-      "password": "pbkdf2_sha256$600000$GvQFwzzAc7tvuZGIigqbLc$VOhEYdnL8xi4uWo9RMZNmLKqLgeIhVaTJPnpw6Fq5ZU=",
-      "last_login": null,
-      "is_superuser": false,
-      "username": "ivanka_new",
-      "first_name": "Ivanka",
-      "last_name": "Kozak",
-      "email": "kozakivanna0@gmail.com",
-      "is_staff": false,
-      "is_active": true,
-      "date_joined": "2023-12-17T13:01:51.986Z",
-      "image_url": "None",
-      "score": 0,
-      "count_tasks": 0,
-      "photo": "None",
-      "groups": [],
-      "user_permissions": []
-    }
-  },
-  {
-    "model": "User.user",
-    "pk": 19,
-    "fields": {
-      "password": "pbkdf2_sha256$600000$2SbPRgmCroUuzXo441NgTw$BzdZ6CWUINEVz98i+81vopU1tFlyWNirs0MPDMHkn8w=",
-      "last_login": null,
-      "is_superuser": false,
-      "username": "test_ivanka",
-      "first_name": "Ivanka",
-      "last_name": "Kozak",
-      "email": "Ivanka.kozak3@gmail.com",
-      "is_staff": false,
-      "is_active": true,
-      "date_joined": "2023-12-17T19:39:40.599Z",
-      "image_url": "None",
-      "score": 4,
-      "count_tasks": 0,
-      "photo": "None",
-      "groups": [],
-      "user_permissions": []
-    }
-  },
-  {
-    "model": "User.user",
-    "pk": 8,
-    "fields": {
-      "password": "pbkdf2_sha256$600000$fyn0zLb2h3DJZoXJamMjQR$Ztkc7SIrnVAwYuJM5suKRAUhraOA33+mWjHDXSff594=",
-      "last_login": null,
-      "is_superuser": false,
-      "username": "test",
-      "first_name": "Test",
-      "last_name": "Test",
-      "email": "ivanna.kozak.shi.2022@lpnu.ua",
-      "is_staff": false,
-      "is_active": true,
-      "date_joined": "2023-12-17T13:15:53.903Z",
-      "image_url": "None",
-      "score": 0,
-      "count_tasks": 0,
-      "photo": "None",
-      "groups": [],
-      "user_permissions": []
-    }
-  },
-  {
-    "model": "User.user",
-    "pk": 16,
-    "fields": {
-      "password": "pbkdf2_sha256$600000$d7B2Q07WcTiW4m8TAJdPCS$E8j0GYfcI1wqrHF81ivSFutYnSO1tyLUIu+mIU6wtmI=",
-      "last_login": null,
-      "is_superuser": false,
-      "username": "Oskar",
-      "first_name": "Oskar",
-      "last_name": "Kobel",
-      "email": "oscgem29@gmail.com",
-      "is_staff": false,
-      "is_active": true,
-      "date_joined": "2023-12-17T18:47:40.790Z",
-      "image_url": "None",
-      "score": 0,
-      "count_tasks": 0,
-      "photo": "None",
-      "groups": [],
-      "user_permissions": []
-    }
-  }
-]
