--- fixtures/usertheme_fixture.json
@@ -1,26 +1,0 @@
-[
-  {
-    "model": "Task.usertheme",
-    "pk": 4,
-    "fields": {
-      "user": "test_ivanka",
-      "theme": "Площа фігур"
-    }
-  },
-  {
-    "model": "Task.usertheme",
-    "pk": 5,
-    "fields": {
-      "user": "test_ivanka",
-      "theme": "Периметр фігур"
-    }
-  },
-  {
-    "model": "Task.usertheme",
-    "pk": 6,
-    "fields": {
-      "user": "test_ivanka",
-      "theme": "На логіку"
-    }
-  }
-]
