+++ requirements.txt
@@ -7,17 +7,12 @@ django-templated-mail==1.1.1
 djoser==2.2.0
 djangorestframework-simplejwt==5.3.0
 django-apscheduler==0.6.2
-django-celery-beat==2.5.0
-django-celery-results==2.5.1
 
 
     # Database Clients
 mysqlclient==2.2.0
 psycopg==3.1.12
 psycopg2==2.9.9
-redis==5.0.1
-flower==2.0.1
-
 
 
     # Authentication and Social Login
@@ -50,7 +45,6 @@ urllib3==2.0.7
 psycopg==3.1.12
 psycopg2==2.9.9
 Pillow==10.1.0
-celery==5.3.6
 
 
     # Timezone Data
